Split middlewares in files, rename worker entry
