+++ packages/cli/src/lib/vite/worker-entry.ts
